毎回テストデータを登録していたのを削除

+++ app/BootStrap.java
@@ -1,5 +1,7 @@
 import java.io.FileNotFoundException;
 
+import models.Event;
+
 import play.jobs.Job;
 import play.jobs.OnApplicationStart;
 import play.test.Fixtures;
@@ -10,9 +12,12 @@ public class BootStrap extends Job {
 	@Override
 	public void doJob() throws FileNotFoundException {
 
-		Fixtures.delete();
+		if(Event.findAll() == null
+				|| Event.findAll().size() == 0){
+			Fixtures.delete();
+			Fixtures.loadModels("initial-data.yml");
 
-		Fixtures.loadModels("initial-data.yml");
+		}
 
 	}
